+++ .gitignore
@@ -1,4 +1,3 @@
-car_racer_gym
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
