--- haptic/videoo/openaigym.episode_batch.0.90203.stats.json
@@ -1,1 +1,0 @@
-{"initial_reset_timestamp": 7423387.088931923, "timestamps": [7423406.424557963], "episode_lengths": [1000], "episode_rewards": [61.40350877193379], "episode_types": ["t"]}
