--- haptic/videoo/openaigym.manifest.0.90203.manifest.json
@@ -1,1 +1,0 @@
-{"stats": "openaigym.episode_batch.0.90203.stats.json", "videos": [["openaigym.video.0.90203.video000000.mp4", "openaigym.video.0.90203.video000000.meta.json"]], "env_info": {"gym_version": "0.21.0"}}
