+++ pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools", "setuptools-scm"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "haptic"
+version = "1.0.0"
+
+[project.scripts]
+haptic = "haptic.__main__:main"
+
+[tool.pytest.ini_options]
+pythonpath = [
+  "src"
+]
