+++ pytest.ini
@@ -1,2 +1,7 @@
+; more suitable for flat-layout
+; [pytest]
+; site_dirs = ./haptic
+
+; more suitable for src-layout
 [pytest]
-site_dirs = ./haptic
+pythonpath = src
